fix: move AgateTable equality helpers out of table.rs
CodeScene flagged table.rs for file size and function count after
__eq__/__ne__ landed there. Keep the Object dispatch thin and put
comparison plus tests in table_eq.rs like join.rs.
